+++ sdk/nodejs/compute/getBackendService.ts
@@ -11,11 +11,13 @@ import * as utilities from "../utilities";
  *
  * > This content is derived from https://github.com/terraform-providers/terraform-provider-google/blob/master/website/docs/d/compute_backend_service.html.markdown.
  */
-export function getBackendService(args: GetBackendServiceArgs, opts?: pulumi.InvokeOptions): Promise<GetBackendServiceResult> {
-    return pulumi.runtime.invoke("gcp:compute/getBackendService:getBackendService", {
+export function getBackendService(args: GetBackendServiceArgs, opts?: pulumi.InvokeOptions): Promise<GetBackendServiceResult> & GetBackendServiceResult {
+    const promise: Promise<GetBackendServiceResult> = pulumi.runtime.invoke("gcp:compute/getBackendService:getBackendService", {
         "name": args.name,
         "project": args.project,
     }, opts);
+
+    return pulumi.utils.liftProperties(promise, opts);
 }
 
 /**