+++ sdk/nodejs/compute/getInstanceGroup.ts
@@ -21,14 +21,16 @@ import * as utilities from "../utilities";
  *
  * > This content is derived from https://github.com/terraform-providers/terraform-provider-google/blob/master/website/docs/d/compute_instance_group.html.markdown.
  */
-export function getInstanceGroup(args?: GetInstanceGroupArgs, opts?: pulumi.InvokeOptions): Promise<GetInstanceGroupResult> {
+export function getInstanceGroup(args?: GetInstanceGroupArgs, opts?: pulumi.InvokeOptions): Promise<GetInstanceGroupResult> & GetInstanceGroupResult {
     args = args || {};
-    return pulumi.runtime.invoke("gcp:compute/getInstanceGroup:getInstanceGroup", {
+    const promise: Promise<GetInstanceGroupResult> = pulumi.runtime.invoke("gcp:compute/getInstanceGroup:getInstanceGroup", {
         "name": args.name,
         "project": args.project,
         "selfLink": args.selfLink,
         "zone": args.zone,
     }, opts);
+
+    return pulumi.utils.liftProperties(promise, opts);
 }
 
 /**