+++ sdk/nodejs/container/getEngineVersions.ts
@@ -36,15 +36,17 @@ import * as utilities from "../utilities";
  *
  * > This content is derived from https://github.com/terraform-providers/terraform-provider-google/blob/master/website/docs/d/container_engine_versions.html.markdown.
  */
-export function getEngineVersions(args?: GetEngineVersionsArgs, opts?: pulumi.InvokeOptions): Promise<GetEngineVersionsResult> {
+export function getEngineVersions(args?: GetEngineVersionsArgs, opts?: pulumi.InvokeOptions): Promise<GetEngineVersionsResult> & GetEngineVersionsResult {
     args = args || {};
-    return pulumi.runtime.invoke("gcp:container/getEngineVersions:getEngineVersions", {
+    const promise: Promise<GetEngineVersionsResult> = pulumi.runtime.invoke("gcp:container/getEngineVersions:getEngineVersions", {
         "location": args.location,
         "project": args.project,
         "region": args.region,
         "versionPrefix": args.versionPrefix,
         "zone": args.zone,
     }, opts);
+
+    return pulumi.utils.liftProperties(promise, opts);
 }
 
 /**