+++ sdk/nodejs/container/getRegistryImage.ts
@@ -24,14 +24,16 @@ import * as utilities from "../utilities";
  *
  * > This content is derived from https://github.com/terraform-providers/terraform-provider-google/blob/master/website/docs/d/container_registry_image.html.markdown.
  */
-export function getRegistryImage(args: GetRegistryImageArgs, opts?: pulumi.InvokeOptions): Promise<GetRegistryImageResult> {
-    return pulumi.runtime.invoke("gcp:container/getRegistryImage:getRegistryImage", {
+export function getRegistryImage(args: GetRegistryImageArgs, opts?: pulumi.InvokeOptions): Promise<GetRegistryImageResult> & GetRegistryImageResult {
+    const promise: Promise<GetRegistryImageResult> = pulumi.runtime.invoke("gcp:container/getRegistryImage:getRegistryImage", {
         "digest": args.digest,
         "name": args.name,
         "project": args.project,
         "region": args.region,
         "tag": args.tag,
     }, opts);
+
+    return pulumi.utils.liftProperties(promise, opts);
 }
 
 /**