+++ sdk/nodejs/projects/getOrganizationPolicy.ts
@@ -25,11 +25,13 @@ import * as utilities from "../utilities";
  *
  * > This content is derived from https://github.com/terraform-providers/terraform-provider-google/blob/master/website/docs/d/project_organization_policy.html.markdown.
  */
-export function getOrganizationPolicy(args: GetOrganizationPolicyArgs, opts?: pulumi.InvokeOptions): Promise<GetOrganizationPolicyResult> {
-    return pulumi.runtime.invoke("gcp:projects/getOrganizationPolicy:getOrganizationPolicy", {
+export function getOrganizationPolicy(args: GetOrganizationPolicyArgs, opts?: pulumi.InvokeOptions): Promise<GetOrganizationPolicyResult> & GetOrganizationPolicyResult {
+    const promise: Promise<GetOrganizationPolicyResult> = pulumi.runtime.invoke("gcp:projects/getOrganizationPolicy:getOrganizationPolicy", {
         "constraint": args.constraint,
         "project": args.project,
     }, opts);
+
+    return pulumi.utils.liftProperties(promise, opts);
 }
 
 /**