+++ sdk/nodejs/serviceAccount/getAccountAccessToken.ts
@@ -12,13 +12,15 @@ import * as utilities from "../utilities";
  *
  * > This content is derived from https://github.com/terraform-providers/terraform-provider-google/blob/master/website/docs/d/service_account_access_token.html.markdown.
  */
-export function getAccountAccessToken(args: GetAccountAccessTokenArgs, opts?: pulumi.InvokeOptions): Promise<GetAccountAccessTokenResult> {
-    return pulumi.runtime.invoke("gcp:serviceAccount/getAccountAccessToken:getAccountAccessToken", {
+export function getAccountAccessToken(args: GetAccountAccessTokenArgs, opts?: pulumi.InvokeOptions): Promise<GetAccountAccessTokenResult> & GetAccountAccessTokenResult {
+    const promise: Promise<GetAccountAccessTokenResult> = pulumi.runtime.invoke("gcp:serviceAccount/getAccountAccessToken:getAccountAccessToken", {
         "delegates": args.delegates,
         "lifetime": args.lifetime,
         "scopes": args.scopes,
         "targetServiceAccount": args.targetServiceAccount,
     }, opts);
+
+    return pulumi.utils.liftProperties(promise, opts);
 }
 
 /**