+++ sdk/nodejs/serviceAccount/getAccountKey.ts
@@ -28,12 +28,14 @@ import * as utilities from "../utilities";
  *
  * > This content is derived from https://github.com/terraform-providers/terraform-provider-google/blob/master/website/docs/d/service_account_key.html.markdown.
  */
-export function getAccountKey(args: GetAccountKeyArgs, opts?: pulumi.InvokeOptions): Promise<GetAccountKeyResult> {
-    return pulumi.runtime.invoke("gcp:serviceAccount/getAccountKey:getAccountKey", {
+export function getAccountKey(args: GetAccountKeyArgs, opts?: pulumi.InvokeOptions): Promise<GetAccountKeyResult> & GetAccountKeyResult {
+    const promise: Promise<GetAccountKeyResult> = pulumi.runtime.invoke("gcp:serviceAccount/getAccountKey:getAccountKey", {
         "name": args.name,
         "project": args.project,
         "publicKeyType": args.publicKeyType,
     }, opts);
+
+    return pulumi.utils.liftProperties(promise, opts);
 }
 
 /**